Implement the gear page styling
I used grid here instead of flexbox to meet one of the project's
requirements.

Currently, only the USA pair of BOA 1" Elite Split Shorts is shown. This
needs to be refactored and one of the six BOA 1" Elite Split Shorts
images should be randomly chosen each time the page is loaded.