Refactor the form labels on pace calculator page
I did not want the form labels to be seen on the page. Previously I had
the input elements as children of the labels and did not provide text
for the labels. Also, a singular label was used for multiple input
elements.

Now, each input element has a label element and the label elements are
hidden by setting display to none.