Clear mile pace output when form is reset
When implementing this, I also realized that the mile pace output
display should have no text by default. It would start out as "Minutes
Per Mile", but would then disappear when the form resets because the
output was cleared. Now, "Minutes Per Mile" is only displayed when the
output is shown.

Resolves: #23